% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lmmcols.R
\name{add_lmmcols}
\alias{add_lmmcols}
\title{Add new columns/varaibles for LMM analysis}
\usage{
add_lmmcols(.data, object, Nstr = 3)
}
\arguments{
\item{.data}{a data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from \code{dbplyr} or \code{dtplyr}). See \code{data} in
\link{model.matrix}.}

\item{object}{an object of an appropriate class. For the default method, a
model \link{formula} or a \link{terms} object. See \code{object} in \link{model.matrix}.}

\item{Nstr}{the number (integer) of strings to be kept in the name.
Default to 3.}
}
\value{
A data frame with new columns/variables.
}
\description{
Based on contrast coding of each independent variable to add new
columns/variables to facilitate linear mixed-effects analysis.
}
\examples{
add_lmmcols(jin2022noncon, isSame ~ Congruency * Alignment + SD) |> str()

}
