% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equi_cor.R
\name{equi_cor}
\alias{equi_cor}
\title{Perform Equivalence Tests for correlations}
\usage{
equi_cor(.data, low_eqbound_r = -1, high_eqbound_r = 1, alpha = 0.1, ...)
}
\arguments{
\item{.data}{a data frame for correlation. Each row is one participant and
each column is one variable.}

\item{low_eqbound_r}{the lower boundary to be tested against. Default to -1.}

\item{high_eqbound_r}{the upper boundary to be tested against. Default to 1.}

\item{alpha}{alpha level of confidence intervals. Default to .10.}

\item{...}{argument pass to \code{\link[TOSTER:TOSTr]{TOSTER::TOSTr()}}.}
}
\value{
a data frame:
\itemize{
\item r the correlation coefficient
\item tost_pmax the maximum p-value of the two one-sided tests.
\item tost_lw, tost_up the boundaries of the confidence intervals for the TOST.
\item alpha alpha level of confidence intervals.
\item low_eqboundar_r, high_eqbound_r the applied boundaries to be tested against with.
}
}
\description{
Perform equivalence tests with \code{\link[TOSTER:TOSTr]{TOSTER::TOSTr()}}, where the
correlation coefficient is obtained from \code{\link[psych:corr.test]{psych::corr.test()}}.
}
\examples{
equi_cor(attitude, -.25, .25)
}
