% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equi_plot.R
\name{equi_plot}
\alias{equi_plot}
\title{Plot hypothetical scenarios for NHST and Equivalence Tests}
\usage{
equi_plot(
  delta_null,
  linecustom = NULL,
  xrange = NULL,
  StrScenario = "Scenario",
  linecolor = c("#D55E00", "#56B4E9", "gray30"),
  jitters = 0.2,
  label = "Correlation coefficient",
  seed = 1215
)
}
\arguments{
\item{delta_null}{one value or a vector of two values. One value will be used
for one-sided tests and two values will be used as two-sided tests.}

\item{linecustom}{a custom group of situations to be displayed.}

\item{xrange}{the range of the x-axis.}

\item{StrScenario}{the string to be displayed on y-axis. Default to "Scenario".}

\item{linecolor}{the three colors to be used for plotting NHST significant,
ET significant, and inconclusive results. Default to red, blue, gray.}

\item{jitters}{some jitters to show the intervals. Default to .2.}

\item{label}{label of x-axis. Default to "Correlation coefficient".}

\item{seed}{seed for generating the jitters.}
}
\value{
a plot (so far).
}
\description{
Plot the hypothetical scenarios for NHST and Equivalence Test.
}
\examples{
equi_plot(c(-.15, .15))
equi_plot(-.3)
equi_plot(.2)
linecustom <- matrix(c(0.10, -0.23, -0.13, -0.04, -0.19, -0.17,
                       0.30, -0.03,  0.07,  0.16,  0.01,  0.17), nrow=2)
equi_plot(c(-.15, .15), linecustom=linecustom)
}
