% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.R
\name{lmerf}
\alias{lmerf}
\title{A wrapper function to save the \code{\link[=lmer]{lmer()}} and \code{\link[=glmer]{glmer()}} fit/output as a local
file, which will be reloaded when the code is re-run.}
\usage{
lmerf(..., file = "none")
}
\arguments{
\item{...}{arguments in \code{\link[=lmer]{lmer()}}}

\item{file}{file name of the local file to be saved. Default to "none", where
no files will be saved/loaded.}
}
\value{
the fit/output.
}
\description{
A wrapper function to save the \code{\link[=lmer]{lmer()}} and \code{\link[=glmer]{glmer()}} fit/output as a local
file, which will be reloaded when the code is re-run.
}
\examples{
fit <- lmerf(log(RT) ~ Congruency * Alignment + (1|SubjID), jin2022noncon, file = "none")
}
