% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkrandom.R
\name{mkrandom}
\alias{mkrandom}
\title{Make random effect variables/columns}
\usage{
mkrandom(.data, ...)
}
\arguments{
\item{.data}{a data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from \code{dbplyr} or \code{dtplyr}).}

\item{...}{\if{html}{\out{<tidy-select>}} One or more unquoted expressions separated by
commas. Variable names can be used as if they were positions in the data
frame, so expressions like x:y can be used to select a range of variables.
see ... in \code{\link[dplyr:select]{dplyr::select()}}.}
}
\value{
a character/string.
}
\description{
Make a string/character to be used as random effect in linear mixed-effect
analysis.
}
\examples{
mkrandom(jin2022noncon, Congruency:SD)
}
