% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotroc.R
\name{plotroc}
\alias{plotroc}
\title{Plot Receiver Operating Characteristic (ROC) curve}
\usage{
plotroc(d = c(1, 2, 3))
}
\arguments{
\item{d}{a vector of sensitivity d' (default to c(1,2,3)).}
}
\value{
a ROC curve plot.
}
\description{
Plot Receiver Operating Characteristic (ROC) curve
}
\examples{
plotroc()
}
