% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsdt.R
\name{plotsdt}
\alias{plotsdt}
\title{Plot Signal Detection Theory (SDT) functions}
\usage{
plotsdt(
  d = 2,
  c_x = d/2,
  xlim = c(-4, d + 4),
  ymax = NULL,
  hit = NULL,
  fa = NULL
)
}
\arguments{
\item{d}{sensitivity d' (default to 2).}

\item{c_x}{criterion on x-axis (it is neither beta or C). It is simply where the criteria on x-axis. Default to d/2.}

\item{xlim}{x-axis limits. Default to c(-4, 6).}

\item{ymax}{y-axis limits. Default to .45.}

\item{hit}{hit rate. Default to NULL. If both hit and fa are provided, d is calculated as qnorm(hit) - qnorm(fa).}

\item{fa}{false alarm rate. Default to NULL. If both hit and fa are provided, d is calculated as qnorm(hit) - qnorm(fa).}
}
\value{
a list with three elements: \code{plot} (SDT plot),  \code{dv} (sensitivity, beta, and C) and \code{rates} (hit, miss, false alarm, and correct rejection).
}
\description{
Plot Signal Detection Theory (SDT) functions
}
\examples{
sdt <- plotsdt()
}
