% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_cor.R
\name{psych_cor}
\alias{psych_cor}
\title{Tidied up correlation outputs}
\usage{
psych_cor(.data, alpha = 0.05, ...)
}
\arguments{
\item{.data}{a data frame for correlation. Each row is one participant and
each column is one variable.}

\item{alpha}{alpha level of confidence intervals. Default to .05.}

\item{...}{arguments pass to \code{psych::corr.test()}.}
}
\value{
data frame of correlations
\itemize{
\item lower, upper: lower and upper of the (uncorrected) confidence interval boundaries
\item r: correlation coefficients
\item p: uncorrected p-value
\item N: sample size
\item alpha: alpha used to calculated the uncorrected CIs
\item p.adj: corrected (adjusted) p-value
\item adjust: the adjust method
\item alt: two-sided tests
\item lower.adj, upper.adj: lower and upper of the corrected confidence interval boundaries
}
}
\description{
Tidied up the output of \code{psych::corr.test()}.
}
\examples{
psych_cor(attitude)
}
