% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2ci.R
\name{r2ci}
\alias{r2ci}
\title{Wrapper to obtain CIs from correlation coefficient.}
\usage{
r2ci(r, n)
}
\arguments{
\item{r}{A correlation coefficient (more see \code{\link[psych:corr.test]{psych::corr.test()}})}

\item{n}{Number of observations (more see \code{\link[psych:corr.test]{psych::corr.test()}})}
}
\value{
the CI
}
\description{
Wrapper to obtain CIs from correlation coefficient.
}
\examples{
r2ci(0.5, 50)
}
