% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_each.R
\name{rel_each}
\alias{rel_each}
\title{Calculate reliability for long-format data frame}
\usage{
rel_each(
  .data,
  DV,
  cor_method = cor,
  item = "test_item",
  fake_tn = FALSE,
  SubjID = "SubjID",
  ...
)
}
\arguments{
\item{.data}{\emph{df} data frame (long format)}

\item{DV}{\emph{str} name of the column for the dependent variable.}

\item{cor_method}{\emph{function handle} the correlation methods to be used. Options
are \code{cor} (default), \code{cov}, and \code{pol} (for \code{\link[psych:tetrachor]{psych::polychoric()}}). When using
\code{cov}, you may also want to specify \code{covar=TRUE}.}

\item{item}{\emph{str} name of the column for testing item (e.g., stimulus, survey
questions). Default to "test_item".}

\item{fake_tn}{\emph{boo} whether to add fake trial indices. Default to False.}

\item{SubjID}{\emph{str} name of the column for subject IDs. Default to "SubjID".}

\item{...}{other arguments pass to \code{\link[psych:guttman]{psych::splitHalf()}}.}
}
\value{
object from \code{\link[psych:guttman]{psych::splitHalf()}}.
}
\description{
\strong{This function are not developed very well! Please use with caution.}
Calculate reliability of data frame in long-format with \code{\link[psych:guttman]{psych::splitHalf()}}.
}
\examples{
jin2022noncon |>
  dplyr::filter(Congruency=="con", Alignment=="ali") |>
  rel_each(DV="RT", fake_tn=TRUE)

}
