% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt.R
\name{sdt}
\alias{sdt}
\title{Calculate d' in Signal Detection Theory}
\usage{
sdt(
  .data,
  SN = "SD",
  isSignal = "isSame",
  SubjID = "SubjID",
  group_other = NULL,
  signal = "same",
  d_correction = "MK1985"
)
}
\arguments{
\item{.data}{\emph{df} dataframe (long format)}

\item{SN}{\emph{str} name of the column in \code{.data} describing whether each trial
showed \emph{signal} or \emph{noise}. Default to "SD".}

\item{isSignal}{\emph{str} name of the column in \code{.data} describing whether it
was reported \emph{signal} on each trial. Default to "isSame".}

\item{SubjID}{\emph{str} name of the column in \code{.data} describing participant IDs.
Default to "SubjID".}

\item{group_other}{\emph{vector of str} a list of columns in \code{.data} to be grouped
by before calculating d'.}

\item{signal}{\emph{str} the level name for \emph{signal} in \code{SN} of \code{.data}. Default
to "same".}

\item{d_correction}{\emph{str} method used to correct 1 and 0 in hits and false
alarm rates. There are two methods:
\itemize{
\item "MK1985" (default): correct 1 and 0 to \verb{(2N-1)/(2N)} and \verb{1/(2N)} where \code{N} is the number of trials in each condition.
\item "SC1988": add 0.5 to number of hits and false alarm and add 1 to the number of trials in each condition.
}}
}
\value{
a list
\itemize{
\item \strong{df} a dataframe containing d
\item \strong{correction} the method used to correct 1 and 0.
}
}
\description{
Calculate d' in Signal Detection Theory
}
\examples{
sdt(jin2022noncon)
sdt(jin2022noncon, d_correction="SC1988")
sdt(jin2022noncon, group_other=c("Congruency", "Alignment"))
}
\references{
Macmillan, N. A., & Kaplan, H. L. (1985). Detection theory analysis of group data: Estimating sensitivity from average hit and false-alarm rates. Psychological Bulletin, 98, 185–199. https://doi.org/10.1037/0033-2909.98.1.185

Snodgrass, J. G., & Corwin, J. (1988). Pragmatics of measuring recognition memory: Applications to dementia and amnesia. Journal of Experimental Psychology: General, 117(1), 34–50. https://doi.org/10.1037/0096-3445.117.1.34
}
