% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_contr.R
\name{set_contr}
\alias{set_contr}
\title{Apply successive difference contrast coding}
\usage{
set_contr(.data, varnames, contraFUN = MASS::contr.sdif)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from \code{dbplyr} or \code{dtplyr})}

\item{varnames}{the columns to be applied contrasting code (without quotes)}

\item{contraFUN}{the contrast coding function (name) to be used.
Default to \code{MASS::contr.sdif}. Currently popular contrast
coding functions are: \href{default}{MASS::contr.sdif}, \link{contr.sum},
\link{contr.treatment}, \link{contr.SAS}, etc.}
}
\value{
A data frame.
}
\description{
Apply successive difference contrast coding (\code{\link[MASS:contr.sdif]{MASS::contr.sdif()}}) to
variables in data and create variables to be used in steps of obtaining the
optimal mixed-effects models.
}
\examples{
set_contr(jin2022noncon, Congruency) |> str()

set_contr(jin2022noncon, c(Congruency, Alignment)) |> str()

set_contr(jin2022noncon, c(Congruency, Alignment), contr.sum) |> str()

}
