% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sdif.R
\name{set_sdif}
\alias{set_sdif}
\title{Apply successive difference contrast coding}
\usage{
set_sdif(.data, varnames)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from \code{dbplyr} or \code{dtplyr})}

\item{varnames}{the columns to be applied contrasting code (without quotes)}
}
\value{
A data frame.
}
\description{
Apply successive difference contrast coding (\code{MASS::contr.sdif()}) to
variables in data and create variables to be used in steps of obtaining the
optimal mixed-effects models.
}
\examples{
set_sdif(jin2022noncon, Congruency) |> str()
set_sdif(jin2022noncon, c(Congruency, Alignment)) |> str()
}
