% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splithalf_d.R
\name{splithalf_d}
\alias{splithalf_d}
\title{Calculate split half reliability for d'}
\usage{
splithalf_d(
  .data,
  SN = "SD",
  SubjID = "SubjID",
  correction = TRUE,
  startseed = 1215,
  iter = 100,
  cores = 1,
  ...
)
}
\arguments{
\item{.data}{\emph{df} dataframe (long format)}

\item{SN}{\emph{str} name of the column in \code{.data} describing whether each trial
showed \emph{signal} or \emph{noise}. Default to "SD".}

\item{SubjID}{\emph{str} name of the column in \code{.data} describing participant IDs.
Default to "SubjID".}

\item{correction}{\emph{boo} whether to apply correction to the reliability.
Default to TRUE.}

\item{startseed}{\emph{int} integer for the starting seed. Default to 1215.}

\item{iter}{\emph{int} number of iterations for calculating the split half
reliability. Default to 100.}

\item{cores}{\emph{int} number of cores used to run the iterations. Default to 1.}

\item{...}{other arguments pass to \link{sdt}.
\itemize{
\item \strong{isSignal} \emph{str} name of the column in \code{.data} describing whether it
was reported \emph{signal} on each trial. Default to "isSame".
\item \strong{signal} \emph{str} the level name for \emph{signal} in \code{SN} of \code{.data}. Default
to "same".
\item \strong{d_correction} \emph{str} method used to correct 1 and 0 in hits and false
alarm rates. There are two methods: "MK1985" (default) and "SC1988". More see
\link{sdt}.
}}
}
\value{
a data frame
\itemize{
\item \strong{iter} the number of iteraction
\item \strong{alpha} cronbach alpha
\item \strong{alpha_cor} cronbach alpha corrected by Spearman-Brown method
\item \strong{seed} seed used to split the data
}
}
\description{
split the data randomly into two halves and calculate the d'
(with \link{sdt}) for the two subsets. Then calculate the Cronbach's alpha
corrected with Spearman-Brown method.
}
\examples{
splithalf_d(jin2022noncon)
}
