% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/touch.R
\name{touch}
\alias{touch}
\title{Make empty files (touch)}
\usage{
touch(path, pattern = NULL)
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds
to the working directory, \code{\link[=getwd]{getwd()}}. Tilde expansion (see \link{path.expand}) is
performed. Missing values will be ignored. Elements with a marked encoding
will be converted to the native encoding (and if that fails, considered
non-existent). See \code{\link[=list.files]{list.files()}}.}

\item{pattern}{an optional \href{regex}{regular expression}. Only file names which
match the regular expression will be returned.}
}
\value{
NA
}
\description{
Make empty files for uploading online.
}
\examples{
# touch("path_to_a_directory")
}
