% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsdt.R
\name{plotsdt}
\alias{plotsdt}
\title{Plot Signal Detection Theory (SDT) functions}
\usage{
plotsdt(
  d = 2,
  c_x = d/2 - 0.75,
  xlim = c(-4, d + 4),
  ymax = NULL,
  hit = NULL,
  fa = NULL,
  showd = FALSE,
  showc = FALSE,
  showb = FALSE
)
}
\arguments{
\item{d}{sensitivity d' (default to 2).}

\item{c_x}{criterion on x-axis (it is neither beta or C). It is simply where the criteria on x-axis. Default to d/2.}

\item{xlim}{x-axis limits. Default to c(-4, 6).}

\item{ymax}{y-axis limits. Default to .45.}

\item{hit}{hit rate. Default to NULL. If both hit and fa are provided, d is calculated as qnorm(hit) - qnorm(fa).}

\item{fa}{false alarm rate. Default to NULL. If both hit and fa are provided, d is calculated as qnorm(hit) - qnorm(fa).}

\item{showd}{whether show d' value on the plot. Default to FALSE. If TRUE, d' is shown in red. If a color is provided, d' is shown in that color.}

\item{showc}{whether show criterion C on the plot. Default to FALSE. If TRUE, C is shown in darkgreen. If a color is provided, C is shown in that color.}

\item{showb}{whether show beta value on the plot. Default to FALSE. If TRUE, beta is shown in blue. If a color is provided, beta is shown in that color.}
}
\value{
a list with three elements: \code{plot} (SDT plot),  \code{dv} (sensitivity, beta, and C) and \code{rates} (hit, miss, false alarm, and correct rejection).
}
\description{
Plot Signal Detection Theory (SDT) functions
}
\examples{
sdt <- plotsdt()
}
